remove naigation compose
