+++ ThemingCodelab/app/build.gradle
@@ -123,8 +123,6 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.5.1"
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.5.1"
 
-    implementation 'androidx.navigation:navigation-compose:2.5.3'
-
     androidTestImplementation 'androidx.test:rules:1.5.0'
     androidTestImplementation 'androidx.test:runner:1.5.2'
