+++ ThemingCodelab/app/src/main/java/com/example/reply/ui/ReplyApp.kt
@@ -19,23 +19,21 @@ package com.example.reply.ui
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Article
+import androidx.compose.material.icons.filled.Inbox
+import androidx.compose.material.icons.filled.People
+import androidx.compose.material.icons.outlined.ChatBubbleOutline
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
-import com.example.reply.ui.navigation.ReplyNavigationActions
-import com.example.reply.ui.navigation.ReplyRoute
-import com.example.reply.ui.navigation.ReplyTopLevelDestination
-import com.example.reply.ui.navigation.TOP_LEVEL_DESTINATIONS
+import com.example.reply.R
 
 @Composable
 fun ReplyApp(
@@ -43,19 +41,8 @@ fun ReplyApp(
     closeDetailScreen: () -> Unit = {},
     navigateToDetail: (Long) -> Unit
 ) {
-    val navController = rememberNavController()
-    val navigationActions = remember(navController) {
-        ReplyNavigationActions(navController)
-    }
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val selectedDestination =
-        navBackStackEntry?.destination?.route ?: ReplyRoute.INBOX
-
     ReplyAppContent(
         replyHomeUIState = replyHomeUIState,
-        navController = navController,
-        selectedDestination = selectedDestination,
-        navigateToTopLevelDestination = navigationActions::navigateTo,
         closeDetailScreen = closeDetailScreen,
         navigateToDetail = navigateToDetail
     )
@@ -65,29 +52,32 @@ fun ReplyApp(
 fun ReplyAppContent(
     modifier: Modifier = Modifier,
     replyHomeUIState: ReplyHomeUIState,
-    navController: NavHostController,
-    selectedDestination: String,
-    navigateToTopLevelDestination: (ReplyTopLevelDestination) -> Unit,
     closeDetailScreen: () -> Unit,
     navigateToDetail: (Long) -> Unit,
 ) {
+
+    val selectedDestination = remember { mutableStateOf( ReplyRoute.INBOX) }
 
     Column(
         modifier = modifier
             .fillMaxSize()
     ) {
-        ReplyNavHost(
-            navController = navController,
-            replyHomeUIState = replyHomeUIState,
-            closeDetailScreen = closeDetailScreen,
-            navigateToDetail = navigateToDetail,
-            modifier = Modifier.weight(1f),
-        )
+
+        if (selectedDestination.value == ReplyRoute.INBOX) {
+            ReplyInboxScreen(
+                replyHomeUIState = replyHomeUIState,
+                closeDetailScreen = closeDetailScreen,
+                navigateToDetail = navigateToDetail,
+            )
+        } else {
+            EmptyComingSoon()
+        }
+
         NavigationBar(modifier = Modifier.fillMaxWidth()) {
             TOP_LEVEL_DESTINATIONS.forEach { replyDestination ->
                 NavigationBarItem(
-                    selected = selectedDestination == replyDestination.route,
-                    onClick = { navigateToTopLevelDestination(replyDestination) },
+                    selected = selectedDestination.value == replyDestination.route,
+                    onClick = { selectedDestination.value = replyDestination.route },
                     icon = {
                         Icon(
                             imageVector = replyDestination.selectedIcon,
@@ -100,34 +90,44 @@ fun ReplyAppContent(
     }
 }
 
-@Composable
-private fun ReplyNavHost(
-    navController: NavHostController,
-    replyHomeUIState: ReplyHomeUIState,
-    closeDetailScreen: () -> Unit,
-    navigateToDetail: (Long) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    NavHost(
-        modifier = modifier,
-        navController = navController,
-        startDestination = ReplyRoute.INBOX,
-    ) {
-        composable(ReplyRoute.INBOX) {
-            ReplyInboxScreen(
-                replyHomeUIState = replyHomeUIState,
-                closeDetailScreen = closeDetailScreen,
-                navigateToDetail = navigateToDetail,
-            )
-        }
-        composable(ReplyRoute.DM) {
-            EmptyComingSoon()
-        }
-        composable(ReplyRoute.ARTICLES) {
-            EmptyComingSoon()
-        }
-        composable(ReplyRoute.GROUPS) {
-            EmptyComingSoon()
-        }
-    }
+
+object ReplyRoute {
+    const val INBOX = "Inbox"
+    const val ARTICLES = "Articles"
+    const val DM = "DirectMessages"
+    const val GROUPS = "Groups"
 }
+
+data class ReplyTopLevelDestination(
+    val route: String,
+    val selectedIcon: ImageVector,
+    val unselectedIcon: ImageVector,
+    val iconTextId: Int
+)
+
+val TOP_LEVEL_DESTINATIONS = listOf(
+    ReplyTopLevelDestination(
+        route = ReplyRoute.INBOX,
+        selectedIcon = Icons.Default.Inbox,
+        unselectedIcon = Icons.Default.Inbox,
+        iconTextId = R.string.tab_inbox
+    ),
+    ReplyTopLevelDestination(
+        route = ReplyRoute.ARTICLES,
+        selectedIcon = Icons.Default.Article,
+        unselectedIcon = Icons.Default.Article,
+        iconTextId = R.string.tab_article
+    ),
+    ReplyTopLevelDestination(
+        route = ReplyRoute.DM,
+        selectedIcon = Icons.Outlined.ChatBubbleOutline,
+        unselectedIcon = Icons.Outlined.ChatBubbleOutline,
+        iconTextId = R.string.tab_inbox
+    ),
+    ReplyTopLevelDestination(
+        route = ReplyRoute.GROUPS,
+        selectedIcon = Icons.Default.People,
+        unselectedIcon = Icons.Default.People,
+        iconTextId = R.string.tab_article
+    )
+)
