+++ ThemingCodelab/app/src/main/res/values/strings.xml
@@ -26,7 +26,7 @@
     <string name="compose">Compose</string>
 
     <string name="empty_screen_title">Screen under construction</string>
-    <string name="empty_screen_subtitle">This screen is still under construction. This sample will help you learn about adaptive layouts in Jetpack Compose</string>
+    <string name="empty_screen_subtitle">This screen is still under construction</string>
 
     <string name="back_button">Back</string>
     <string name="more_options_button">More options</string>
